[METAL] Fix deferred memory release in Metal runtime
In case when we build runtime without ARC, we can have problems with
memory releasing. Due to some of Objective-C methods returns
autoreleased pointers, we should specify `autoreleasepool` blocks to
determine life cycle of these pointers.

Without this fix I had have a problem that memory was over when a
topology was executed multiple times.